% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_error.R
\name{binom_error}
\alias{binom_error}
\title{Estimates genotype error probabilities using binomial distribution for allele counts}
\usage{
binom_error(vcfR.object = NULL, onemap.object = NULL,
  vcf.par = c("AD", "DPR"), parent1 = "P1", parent2 = "P2",
  recovering = FALSE, mean_phred = 20, depths = NULL)
}
\arguments{
\item{mean_phred}{mean quality phred value of the sequencing}

\item{depth_matrix}{list containing ref allele counts for parents and progeny, output 
from extract_depth function}
}
\value{
onemap object with error updated
}
\description{
Uses allele counts for each genotype from vcf, generated with NGS genotyping, to estimate 
genotype error probability. These probabilities can be used in multipoint approach to 
weighted the markers according with their sequecing quality
}
\references{
Bilton,T. P. , Schofield,  M. R. , Black,  M. A. , Chagné, D. , Wilcox, P. L. 
and Dodds, K. G. (2018) Accounting for Errors in Low Coverage High-Throughput
Sequencing Data When Constructing Genetic Maps Using Biparental Outcrossed Populations.
\emph{Genetics}. ISSN 0016-6731.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
