% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_genotype.R
\name{binom_genotype}
\alias{binom_genotype}
\title{Estimates genotype error probabilities using binomial distribution for allele counts}
\usage{
binom_genotype(
  vcfR.object = NULL,
  onemap.object = NULL,
  vcf.par = c("AD", "DPR"),
  parent1 = "P1",
  parent2 = "P2",
  recovering = FALSE,
  mean_phred = 20,
  depths = NULL
)
}
\arguments{
\item{vcfR.object}{object output from vcfR package}

\item{onemap.object}{onemap object output from read_onemap, read_mapmaker or onemap_read_vcf function}

\item{vcf.par}{vcf format field that contain alelle counts informations, usually AD and DPR}

\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{recovering}{TRUE/FALSE, if TRUE avaliate all markers from vcf file, if FALSE avaliate only markers in onemap object}

\item{mean_phred}{mean quality phred value of the sequencing}

\item{depths}{matrix with allele depths}
}
\value{
onemap object with error updated
}
\description{
Uses allele counts for each genotype from vcf, generated with NGS genotyping, to estimate 
genotype error probability. These probabilities can be used in multipoint approach to 
weighted the markers according with their sequecing quality
}
\references{
Bilton,T. P. , Schofield,  M. R. , Black,  M. A. , Chagné, D. , Wilcox, P. L. 
and Dodds, K. G. (2018) Accounting for Errors in Low Coverage High-Throughput
Sequencing Data When Constructing Genetic Maps Using Biparental Outcrossed Populations.
\emph{Genetics}. ISSN 0016-6731.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
