% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segregation.tests.R
\name{Bonferroni.alpha}
\alias{Bonferroni.alpha}
\title{Bonferroni.alpha}
\usage{
Bonferroni.alpha(x, global.alpha = 0.05)
}
\arguments{
\item{x}{an object of class onemap.segreg.test}

\item{global.alpha}{the global alpha that is target}
}
\value{
the alpha value (numeric)
}
\description{
It shows the alpha value to be used in each test to control global type I error
for chi-square tests for segregation of all markers if Bonferroni's criteria is applied.
}
\examples{
data(fake.bc.onemap) # Loads a fake backcross dataset installed with onemap
Chi <- test.segregation(fake.bc.onemap) # Performs the chi-square test for all markers
print(Chi) # Shows the results of the Chi-square tests
Bonferroni.alpha (Chi) # Shows the global alpha level using Bonferroni's criteria
}

