% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_twopts}
\alias{check_twopts}
\title{Twopts object sanity check}
\usage{
check_twopts(x)
}
\arguments{
\item{x}{an object of class \code{onemap}}
}
\value{
if consistent, returns 0. If not consistent, returns a 
        vector with a number of tests, where \code{TRUE} indicates
        a failed test.
}
\description{
Based on MAPpoly check_data_sanity function by Marcelo Mollinari
}
\examples{

data(onemap_example_bc)
check_twopts(onemap_example_bc)


}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
