% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.raw.data.R
\name{create.dataframe.outcross.plot}
\alias{create.dataframe.outcross.plot}
\title{create.dataframe.outcross.plot}
\usage{
create.dataframe.outcross.plot(x)
}
\arguments{
\item{x}{an object of class outcross, with data and additional information}
}
\value{
a dataframe
}
\description{
An internal function that prepares a dataframe suitable for
drawing a graphic of raw data using ggplot2, i. e., a data frame
with long format
}

