% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bins_by_prob.R
\name{create_data_bins_by_prob}
\alias{create_data_bins_by_prob}
\title{Not ready yet - use carefully}
\usage{
create_data_bins_by_prob(onemap_obj, bins)
}
\description{
Not ready yet - use carefully
}
