% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{create_dataframe_for_plot_outcross}
\alias{create_dataframe_for_plot_outcross}
\title{create_dataframe_for_plot_outcross}
\usage{
create_dataframe_for_plot_outcross(x)
}
\arguments{
\item{x}{an object of class outcross, with data and additional information}
}
\value{
a dataframe
}
\description{
An internal function that prepares a dataframe suitable for
drawing a graphic of raw data using ggplot2, i. e., a data frame
with long format
}

