% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_probs.R
\name{create_probs}
\alias{create_probs}
\title{Converts probabilities to onemap format}
\usage{
create_probs(
  input.obj = NULL,
  global_error = NULL,
  genotypes_errors = NULL,
  genotypes_probs = NULL
)
}
\arguments{
\item{input.obj}{object of class onemap or onemap sequence}

\item{global_error}{single value to be considered as error probability in HMM emission function}

\item{genotypes_errors}{matrix individuals x markers with error values for each marker}

\item{genotypes_probs}{table containing the probability distribution for each combination of marker × individual. 
Each line on this table represents the combination of one marker with one individual, and the respective probabilities.
The table should contain four three columns (prob(AA), prob(AB) and prob(BB)) and individuals x markers rows.}
}
\description{
Converts probabilities to onemap format
}
