% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_geno.R
\name{draw_geno}
\alias{draw_geno}
\title{Draw Genotypes}
\usage{
draw_geno(input.map, ind, most.likely = F, position = "stack",
  main = "Genotypes")
}
\arguments{
\item{input.map}{map(s). Object(s) of class \code{sequence}.}

\item{ind}{Individual number}

\item{most.likely}{TRUE or FALSE}

\item{position}{"split" or "stack"}

\item{main}{an overall title for the plot. Default is \code{NULL}.}
}
\description{
Provides a draw of the individuals genotypes.
}
\examples{

\dontrun{
 data("onemap_example_out")
 twopt <- rf_2pts(onemap_example_out)
 lg <- group(make_seq(twopt, "all"))
 map1 <- map(make_seq(order_seq(input.seq = make_seq(lg, 1), twopt.alg = "rcd"), "force"))
 map2 <- map(make_seq(order_seq(input.seq = make_seq(lg, 2), twopt.alg = "rcd"), "force"))
 map3 <- map(make_seq(order_seq(input.seq = make_seq(lg, 3), twopt.alg = "rcd"), "force"))
 draw_geno(map1, map2, map3, ind = 1)
 
 data("onemap_example_f2")
 twopt <- rf_2pts(onemap_example_f2)
 lg <- group(make_seq(twopt, "all"))
 map <- list(
 map(make_seq(order_seq(input.seq = make_seq(lg, 1),twopt.alg = "rcd"), "force")),
 map(make_seq(order_seq(input.seq = make_seq(lg, 2),twopt.alg = "rcd"), "force")),
 map(make_seq(order_seq(input.seq = make_seq(lg, 3),twopt.alg = "rcd"), "force"))
 )
 draw_geno(map, 1:2)

}
}
\author{
Getulio Caixeta Ferreira, \email{getulio.caifer@gmail.com}

Cristiane Taniguti, \email{chtaniguti@usp.br}
}
\keyword{utilities}
