% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_haplotypes.R
\name{draw_haplotypes}
\alias{draw_haplotypes}
\title{Draw Genotypes}
\usage{
draw_haplotypes(
  ...,
  ind = 1,
  most.likely = FALSE,
  col = NULL,
  position = "stack",
  show.markers = TRUE,
  group.names = NULL,
  main = "Genotypes"
)
}
\arguments{
\item{...}{Map(s) or list(s) of maps. Object(s) of class sequence.}

\item{ind}{The individual number to be ploted.}

\item{most.likely}{logical; if  \code{TRUE}, the most likely genotype is plotted; if FALSE (default) the genotype probability is plotted.}

\item{col}{Color of parentes' homologous.}

\item{position}{"split" or "stack"; if "split" (default) the parents' homologous are plotted separately. if "stack" the parents' homologous are plotted together.}

\item{show.markers}{logical; if  \code{TRUE}, the markers (default) are plotted.}

\item{group.names}{Names of the groups.}

\item{main}{An overall title for the plot; default is \code{NULL}.}
}
\description{
Provides a draw of the individuals genotypes.
}
\examples{

\dontrun{
 data("onemap_example_out")
 twopt <- rf_2pts(onemap_example_out)
 lg <- group(make_seq(twopt, "all"))
 map1 <- map(make_seq(order_seq(input.seq = make_seq(lg, 1), twopt.alg = "rcd"), "force"))
 map2 <- map(make_seq(order_seq(input.seq = make_seq(lg, 2), twopt.alg = "rcd"), "force"))
 map3 <- map(make_seq(order_seq(input.seq = make_seq(lg, 3), twopt.alg = "rcd"), "force"))
 draw_geno(map1, map2, map3, ind = 1)
 
 data("onemap_example_f2")
 twopt <- rf_2pts(onemap_example_f2)
 lg <- group(make_seq(twopt, "all"))
 maps <- list(
 map(make_seq(order_seq(input.seq = make_seq(lg, 1),twopt.alg = "rcd"), "force")),
 map(make_seq(order_seq(input.seq = make_seq(lg, 2),twopt.alg = "rcd"), "force")),
 map(make_seq(order_seq(input.seq = make_seq(lg, 3),twopt.alg = "rcd"), "force"))
 )
 draw_geno(maps, ind= 1:2)

}
}
\author{
Getulio Caixeta Ferreira, \email{getulio.caifer@gmail.com}

Cristiane Taniguti, \email{chtaniguti@usp.br}
}
\keyword{utilities}
