% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{edit_order_onemap}
\alias{edit_order_onemap}
\title{Edit sequence ordered by reference genome positions 
comparing to another set order}
\usage{
edit_order_onemap(input.seq)
}
\arguments{
\item{input.seq}{object of class sequence with alternative order (not genomic order)}
}
\description{
Edit sequence ordered by reference genome positions 
comparing to another set order
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
