% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_onemap_obj}
\alias{empty_onemap_obj}
\title{Produce empty object to avoid code break. Function for internal purpose.}
\usage{
empty_onemap_obj(vcf, P1, P2, cross)
}
\arguments{
\item{vcf}{object of class vcfR}

\item{P1}{character with parent 1 ID}

\item{P2}{character with parent 2 ID}

\item{cross}{type of cross. Must be one of: \code{"outcross"} for full-sibs;
\code{"f2 intercross"} for an F2 intercross progeny; \code{"f2 backcross"};
\code{"ri self"} for recombinant inbred lines by self-mating; or
\code{"ri sib"} for recombinant inbred lines by sib-mating.}
}
\description{
Produce empty object to avoid code break. Function for internal purpose.
}
