% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_hmm_out}
\alias{est_hmm_out}
\title{Run HMM chains}
\usage{
est_hmm_out(geno_R, error_R, type_R, phase_R, rf_R, verbose_R, tol_R)
}
\arguments{
\item{geno_R}{genotypes}

\item{error_R}{genotypes probabilities}

\item{type_R}{marker types}

\item{phase_R}{phase estimated by 2-pts}

\item{rf_R}{recombination fraction}

\item{verbose_R}{logical to display or not the procedure}

\item{tol_R}{EM algorithm tolerance}
}
\value{
vector with recombination fraction estimated by multipoint approach
}
\description{
Run HMM chains
}
