% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{export_mappoly_genoprob}
\alias{export_mappoly_genoprob}
\title{Export genotype probabilities in MAPpoly format (input for QTLpoly)}
\usage{
export_mappoly_genoprob(input.map)
}
\arguments{
\item{input.map}{object of class `sequence`}
}
\value{
object of class `mappoly.genoprob`
}
\description{
Export genotype probabilities in MAPpoly format (input for QTLpoly)
}
