% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{export_viewpoly}
\alias{export_viewpoly}
\title{Export OneMap maps to be visualized in VIEWpoly}
\usage{
export_viewpoly(seqs.list)
}
\arguments{
\item{seqs.list}{a list with `sequence` objects}
}
\value{
object of class viewmap
}
\description{
Export OneMap maps to be visualized in VIEWpoly
}
