% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_depth.R
\name{extract_depth}
\alias{extract_depth}
\title{Extract allele counts of progeny and parents of vcf file}
\usage{
extract_depth(vcfR.object = NULL, onemap.object = NULL,
  vcf.par = c("GQ", "AD", "DPR"), parent1 = "P1", parent2 = "P2",
  f1 = "F1", mean_phred = 20, recovering = FALSE)
}
\arguments{
\item{vcfR.object}{object output from vcfR package}

\item{onemap.object}{onemap object output from read_onemap, read_mapmaker or onemap_read_vcf function}

\item{vcf.par}{vcf format field that contain alelle counts informations, usually AD and DPR}

\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{recovering}{TRUE/FALSE, if TRUE avaliate all markers from vcf file, if FALSE avaliate only markers in onemap object}
}
\value{
list containing the following components: \item{palt}{a \code{matrix} with parent 1 and 2 
alternative alelle counts.} \item{pref}{a \code{matrix} with parent 1 and 2 
reference alelle counts.} \item{psize}{a \code{matrix} with parent 1 and 2 
total alelle counts.}\item{oalt}{a \code{matrix} with progeny 
alternative alelle counts.}\item{oref}{a \code{matrix} with progeny 
reference alelle counts.}\item{osize}{a \code{matrix} with progeny 
total alelle counts.}\item{n.mks}{total number of markers.} 
\item{n.ind}{total number of individuals in progeny.} \item{inds}{progeny individuals identification.}
\item{mks}{markers identification.} \item{onemap.object}{same onemap.object inputed}
}
\description{
Uses vcfR package and onemap object to generates list of vectors with
reference alelle count and total counts for each marker and genotypes 
included in onemap object
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
