% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_missing}
\alias{filter_missing}
\title{Filter markers according with a missing data threshold}
\usage{
filter_missing(onemap.obj = NULL, threshold = 0.25)
}
\arguments{
\item{onemap.obj}{an object of class \code{onemap}.}

\item{threshold}{a numeric from 0 to 1 to define the threshold of missing data allowed}
}
\value{
Returns an object of class \code{onemap}
}
\description{
Filter markers according with a missing data threshold
}
\examples{

  data(onemap_example_out)
  filt_obj <- filter_missing(onemap_example_out, threshold=0.25)
  
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
