% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_prob}
\alias{filter_prob}
\title{Function filter genotypes by genotype probability}
\usage{
filter_prob(onemap.obj = NULL, threshold = 0.8)
}
\arguments{
\item{onemap.obj}{an object of class \code{onemap}.}

\item{threshold}{a numeric from 0 to 1 to define the threshold for 
the probability of the called genotype (highest probability)}
}
\value{
Returns an object of class \code{onemap}
}
\description{
Function filter genotypes by genotype probability
}
\examples{

  data(onemap_example_out)
  filt_obj <- filter_prob(onemap_example_out, threshold=0.8)


}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
