% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_batches.R
\name{generate_overlapping_batches}
\alias{generate_overlapping_batches}
\title{Function to divide the sequence in batches with user defined size}
\usage{
generate_overlapping_batches(input.seq, size = 50, overlap = 15)
}
\arguments{
\item{input.seq}{an object of class \code{sequence}.}

\item{size}{The center size around which an optimum is to be searched}

\item{overlap}{The desired overlap between batches}
}
\description{
Function to divide the sequence in batches with user defined size
}
