% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_func.R
\name{haldane}
\alias{haldane}
\title{Apply Haldane mapping function}
\usage{
haldane(rcmb)
}
\arguments{
\item{rcmb}{vector of recombination fraction values}
}
\value{
vector with centimorgan values
}
\description{
Apply Haldane mapping function
}
