% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_only_selected_mks}
\alias{keep_only_selected_mks}
\title{Keep in the onemap and twopts object only markers in the sequences}
\usage{
keep_only_selected_mks(list.sequences = NULL)
}
\arguments{
\item{list.sequences}{a list of objects 'sequence'}
}
\value{
a list of objects 'sequences' with internal onemap and twopts objects reduced
}
\description{
Keep in the onemap and twopts object only markers in the sequences
}
\author{
Cristiane Taniguti
}
