% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_func.R
\name{kosambi}
\alias{kosambi}
\title{Apply kosambi function to recombination fraction values}
\usage{
kosambi(rcmb)
}
\description{
@param rcmb vector with recombination fraction values
 
 @return vector with centimorgan values
}
