\name{LG1_f2_ord}
\alias{LG1_f2_ord}
\docType{data}
\title{
Dataset needed for build Inbred vignette. It is the result from order_seq function exemplified there.
}
\description{
Dataset needed for build Inbred vignette. It is the result from order_seq function exemplified there.
}
\usage{data("LG1_f2_ord")}
\format{
  The format is:
List of 9
 $ ord      :List of 7
  ..$ seq.num   : int [1:17] 89 33 90 37 8 51 5 25 61 66 ...
  ..$ seq.phases: int [1:16] 1 1 1 2 3 1 4 1 1 1 ...
  ..$ seq.rf    : num [1:16] 0.0551 0.0144 0.0632 0.0579 0.1151 ...
  ..$ seq.like  : num -1232
  ..$ data.name :List of 11
  .. ..$ geno         : num [1:200, 1:90] 1 3 2 2 1 0 3 1 1 3 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:200] "IND1" "IND2" "IND3" "IND4" ...
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..$ n.ind        : int 200
  .. ..$ n.mar        : int 90
  .. ..$ segr.type    : chr [1:90] "A.H.B" "C.A" "D.B" "C.A" ...
  .. ..$ segr.type.num: num [1:90] 4 7 6 7 7 6 4 7 6 4 ...
  .. ..$ n.phe        : num 1
  .. ..$ pheno        : num [1:200, 1] 37.6 36.4 37.2 35.8 37.1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr "Trait_1"
  .. ..$ CHROM        : chr [1:90] NA NA NA NA ...
  .. ..$ POS          : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ input        : NULL
  .. ..$ error        : num [1:18000, 1:4] 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:18000] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. .. ..$ : NULL
  .. ..- attr(*, "class")= chr [1:2] "onemap" "f2"
  ..$ probs     : num [1:4, 1:3400] 0.1075 0.7569 0.028 0.1075 0.0701 ...
  ..$ twopt     :List of 8
  .. ..$ data.name:List of 11
  .. .. ..$ geno         : num [1:200, 1:90] 1 3 2 2 1 0 3 1 1 3 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:200] "IND1" "IND2" "IND3" "IND4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ n.ind        : int 200
  .. .. ..$ n.mar        : int 90
  .. .. ..$ segr.type    : chr [1:90] "A.H.B" "C.A" "D.B" "C.A" ...
  .. .. ..$ segr.type.num: num [1:90] 4 7 6 7 7 6 4 7 6 4 ...
  .. .. ..$ n.phe        : num 1
  .. .. ..$ pheno        : num [1:200, 1] 37.6 36.4 37.2 35.8 37.1 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr "Trait_1"
  .. .. ..$ CHROM        : chr [1:90] NA NA NA NA ...
  .. .. ..$ POS          : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  .. .. ..$ input        : NULL
  .. .. ..$ error        : num [1:18000, 1:4] 1 1 1 1 1 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:18000] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. .. .. ..$ : NULL
  .. .. ..- attr(*, "class")= chr [1:2] "onemap" "f2"
  .. ..$ n.mar    : int 90
  .. ..$ LOD      : num 3
  .. ..$ max.rf   : num 0.5
  .. ..$ input    : NULL
  .. ..$ CHROM    : chr [1:90] NA NA NA NA ...
  .. ..$ POS      : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ analysis :List of 4
  .. .. ..$ CC: num [1:90, 1:90] 0 0.465 0.162 0.754 0.486 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ CR: num [1:90, 1:90] 0 0.535 0.162 0.246 0.514 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ RC: num [1:90, 1:90] 0 0.465 0.838 0.754 0.486 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ RR: num [1:90, 1:90] 0 0.535 0.838 0.246 0.514 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..- attr(*, "class")= chr [1:2] "rf_2pts" "f2"
  ..- attr(*, "class")= chr "sequence"
 $ mrk.unpos: int 2
 $ LOD.unpos: num [1, 1:18] -33.6 -72 -85.6 -66.6 -56.8 ...
 $ THRES    : num 3
 $ ord.all  :List of 7
  ..$ seq.num   : int [1:18] 89 33 90 37 8 51 5 25 61 66 ...
  ..$ seq.phases: int [1:17] 1 1 1 2 3 1 4 1 1 1 ...
  ..$ seq.rf    : num [1:17] 0.0551 0.0144 0.0632 0.0579 0.1151 ...
  ..$ seq.like  : num -1256
  ..$ data.name :List of 11
  .. ..$ geno         : num [1:200, 1:90] 1 3 2 2 1 0 3 1 1 3 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:200] "IND1" "IND2" "IND3" "IND4" ...
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..$ n.ind        : int 200
  .. ..$ n.mar        : int 90
  .. ..$ segr.type    : chr [1:90] "A.H.B" "C.A" "D.B" "C.A" ...
  .. ..$ segr.type.num: num [1:90] 4 7 6 7 7 6 4 7 6 4 ...
  .. ..$ n.phe        : num 1
  .. ..$ pheno        : num [1:200, 1] 37.6 36.4 37.2 35.8 37.1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr "Trait_1"
  .. ..$ CHROM        : chr [1:90] NA NA NA NA ...
  .. ..$ POS          : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ input        : NULL
  .. ..$ error        : num [1:18000, 1:4] 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:18000] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. .. ..$ : NULL
  .. ..- attr(*, "class")= chr [1:2] "onemap" "f2"
  ..$ probs     : num [1:4, 1:3600] 0.1075 0.7569 0.028 0.1075 0.0701 ...
  ..$ twopt     :List of 8
  .. ..$ data.name:List of 11
  .. .. ..$ geno         : num [1:200, 1:90] 1 3 2 2 1 0 3 1 1 3 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:200] "IND1" "IND2" "IND3" "IND4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ n.ind        : int 200
  .. .. ..$ n.mar        : int 90
  .. .. ..$ segr.type    : chr [1:90] "A.H.B" "C.A" "D.B" "C.A" ...
  .. .. ..$ segr.type.num: num [1:90] 4 7 6 7 7 6 4 7 6 4 ...
  .. .. ..$ n.phe        : num 1
  .. .. ..$ pheno        : num [1:200, 1] 37.6 36.4 37.2 35.8 37.1 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : NULL
  .. .. .. .. ..$ : chr "Trait_1"
  .. .. ..$ CHROM        : chr [1:90] NA NA NA NA ...
  .. .. ..$ POS          : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  .. .. ..$ input        : NULL
  .. .. ..$ error        : num [1:18000, 1:4] 1 1 1 1 1 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:18000] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. .. .. ..$ : NULL
  .. .. ..- attr(*, "class")= chr [1:2] "onemap" "f2"
  .. ..$ n.mar    : int 90
  .. ..$ LOD      : num 3
  .. ..$ max.rf   : num 0.5
  .. ..$ input    : NULL
  .. ..$ CHROM    : chr [1:90] NA NA NA NA ...
  .. ..$ POS      : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ analysis :List of 4
  .. .. ..$ CC: num [1:90, 1:90] 0 0.465 0.162 0.754 0.486 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ CR: num [1:90, 1:90] 0 0.535 0.162 0.246 0.514 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ RC: num [1:90, 1:90] 0 0.465 0.838 0.754 0.486 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. ..$ RR: num [1:90, 1:90] 0 0.535 0.838 0.246 0.514 ...
  .. .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..- attr(*, "class")= chr [1:2] "rf_2pts" "f2"
  ..- attr(*, "class")= chr "sequence"
 $ data.name:List of 11
  ..$ geno         : num [1:200, 1:90] 1 3 2 2 1 0 3 1 1 3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:200] "IND1" "IND2" "IND3" "IND4" ...
  .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  ..$ n.ind        : int 200
  ..$ n.mar        : int 90
  ..$ segr.type    : chr [1:90] "A.H.B" "C.A" "D.B" "C.A" ...
  ..$ segr.type.num: num [1:90] 4 7 6 7 7 6 4 7 6 4 ...
  ..$ n.phe        : num 1
  ..$ pheno        : num [1:200, 1] 37.6 36.4 37.2 35.8 37.1 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr "Trait_1"
  ..$ CHROM        : chr [1:90] NA NA NA NA ...
  ..$ POS          : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  ..$ input        : NULL
  ..$ error        : num [1:18000, 1:4] 1 1 1 1 1 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:18000] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. ..$ : NULL
  ..- attr(*, "class")= chr [1:2] "onemap" "f2"
 $ probs2   : num [1:4, 1:3400] 0.1075 0.7569 0.028 0.1075 0.0701 ...
 $ probs3   : num [1:4, 1:3600] 0.1075 0.7569 0.028 0.1075 0.0701 ...
 $ twopt    :List of 8
  ..$ data.name:List of 11
  .. ..$ geno         : num [1:200, 1:90] 1 3 2 2 1 0 3 1 1 3 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:200] "IND1" "IND2" "IND3" "IND4" ...
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..$ n.ind        : int 200
  .. ..$ n.mar        : int 90
  .. ..$ segr.type    : chr [1:90] "A.H.B" "C.A" "D.B" "C.A" ...
  .. ..$ segr.type.num: num [1:90] 4 7 6 7 7 6 4 7 6 4 ...
  .. ..$ n.phe        : num 1
  .. ..$ pheno        : num [1:200, 1] 37.6 36.4 37.2 35.8 37.1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr "Trait_1"
  .. ..$ CHROM        : chr [1:90] NA NA NA NA ...
  .. ..$ POS          : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ input        : NULL
  .. ..$ error        : num [1:18000, 1:4] 1 1 1 1 1 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:18000] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. .. ..$ : NULL
  .. ..- attr(*, "class")= chr [1:2] "onemap" "f2"
  ..$ n.mar    : int 90
  ..$ LOD      : num 3
  ..$ max.rf   : num 0.5
  ..$ input    : NULL
  ..$ CHROM    : chr [1:90] NA NA NA NA ...
  ..$ POS      : num [1:90] NA NA NA NA NA NA NA NA NA NA ...
  ..$ analysis :List of 4
  .. ..$ CC: num [1:90, 1:90] 0 0.465 0.162 0.754 0.486 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..$ CR: num [1:90, 1:90] 0 0.535 0.162 0.246 0.514 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..$ RC: num [1:90, 1:90] 0 0.465 0.838 0.754 0.486 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. ..$ RR: num [1:90, 1:90] 0 0.535 0.838 0.246 0.514 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:90] "M1" "M2" "M3" "M4" ...
  ..- attr(*, "class")= chr [1:2] "rf_2pts" "f2"
 - attr(*, "class")= chr "order"
}
\details{
Inbred vignette uses it to save time to build itself.
}
\keyword{datasets}
