\name{LG3_comp}
\alias{LG3_comp}
\docType{data}
\title{
Dataset needed for build Outcrossing vignette. It is the result from compare function exemplified there.
}
\description{
Dataset needed for build Outcrossing vignette. It is the result from compare function exemplified there.
}
\usage{data("LG3_comp")}
\format{
  List of 7
 $ best.ord      : int [1:51, 1:7] 43 43 22 22 7 7 7 7 7 7 ...
 $ best.ord.rf   : num [1:51, 1:6] 0.13186 0.12343 0.12852 0.12034 0.00481 ...
 $ best.ord.phase: int [1:51, 1:6] 1 1 1 1 1 1 1 2 2 1 ...
 $ best.ord.like : num [1:51] -399 -399 -399 -399 -399 ...
 $ best.ord.LOD  : num [1:51] 0 -0.0509 -0.1518 -0.2156 -0.2512 ...
 $ data.name     :List of 11
  ..$ geno         : num [1:100, 1:52] 2 2 2 2 3 2 2 1 2 3 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:100] "IND1" "IND2" "IND3" "IND4" ...
  .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  ..$ n.ind        : int 100
  ..$ n.mar        : int 52
  ..$ segr.type    : chr [1:52] "B3.7" "D2.18" "D1.13" "A.4" ...
  ..$ segr.type.num: num [1:52] 4 7 6 1 7 4 7 4 6 7 ...
  ..$ n.phe        : num 3
  ..$ pheno        : num [1:100, 1:3] 43 12 20 18 34 14 12 53 22 13 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:3] "Pheno1" "Pheno2" "Pheno3"
  ..$ CHROM        : chr [1:52] NA NA NA NA ...
  ..$ POS          : num [1:52] NA NA NA NA NA NA NA NA NA NA ...
  ..$ input        : NULL
  ..$ error        : num [1:5200, 1:4] 5e-06 5e-06 5e-06 1e+00 1e+00 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:5200] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. ..$ : NULL
  ..- attr(*, "class")= chr [1:2] "onemap" "outcross"
 $ twopt         :List of 8
  ..$ data.name:List of 11
  .. ..$ geno         : num [1:100, 1:52] 2 2 2 2 3 2 2 1 2 3 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:100] "IND1" "IND2" "IND3" "IND4" ...
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. ..$ n.ind        : int 100
  .. ..$ n.mar        : int 52
  .. ..$ segr.type    : chr [1:52] "B3.7" "D2.18" "D1.13" "A.4" ...
  .. ..$ segr.type.num: num [1:52] 4 7 6 1 7 4 7 4 6 7 ...
  .. ..$ n.phe        : num 3
  .. ..$ pheno        : num [1:100, 1:3] 43 12 20 18 34 14 12 53 22 13 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : NULL
  .. .. .. ..$ : chr [1:3] "Pheno1" "Pheno2" "Pheno3"
  .. ..$ CHROM        : chr [1:52] NA NA NA NA ...
  .. ..$ POS          : num [1:52] NA NA NA NA NA NA NA NA NA NA ...
  .. ..$ input        : NULL
  .. ..$ error        : num [1:5200, 1:4] 5e-06 5e-06 5e-06 1e+00 1e+00 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:5200] "M1_IND1" "M2_IND1" "M3_IND1" "M4_IND1" ...
  .. .. .. ..$ : NULL
  .. ..- attr(*, "class")= chr [1:2] "onemap" "outcross"
  ..$ n.mar    : int 52
  ..$ LOD      : num 3
  ..$ max.rf   : num 0.5
  ..$ input    : NULL
  ..$ CHROM    : chr [1:52] NA NA NA NA ...
  ..$ POS      : num [1:52] NA NA NA NA NA NA NA NA NA NA ...
  ..$ analysis :List of 4
  .. ..$ CC: num [1:52, 1:52] 0 0.182 0.295 0.575 0.545 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. ..$ CR: num [1:52, 1:52] 0 0.818 0.295 0.509 0.455 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. ..$ RC: num [1:52, 1:52] 0 0.182 0.705 0.491 0.545 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. ..$ RR: num [1:52, 1:52] 0 0.818 0.705 0.425 0.455 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  .. .. .. ..$ : chr [1:52] "M1" "M2" "M3" "M4" ...
  ..- attr(*, "class")= chr [1:2] "rf_2pts" "outcross"
 - attr(*, "class")= chr "compare"
}
\details{
Outcrossing vignette uses it to save time to build itself.
}
\keyword{datasets}
