% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{load_onemap_sequences}
\alias{load_onemap_sequences}
\title{Load list of sequences saved by save_onemap_sequences}
\usage{
load_onemap_sequences(filename)
}
\arguments{
\item{filename}{name of the file to be loaded}
}
\description{
Load list of sequences saved by save_onemap_sequences
}
