% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_window.R
\name{map_window}
\alias{map_window}
\title{Construct the linkage map for a sequence of markers using sliding windows}
\usage{
map_window(input.seq, window = 40, step = 20, tol = 1e-04,
  verbose = FALSE)
}
\arguments{
\item{input.seq}{an object of class \code{sequence}.}

\item{tol}{tolerance for the C routine, i.e., the value used to evaluate
convergence.}

\item{verbose}{If \code{TRUE}, print tracing information.}
}
\value{
An object of class \code{sequence}, which is a list containing the
following components: \item{seq.num}{a \code{vector} containing the
(ordered) indices of markers in the sequence, according to the input file.}
\item{seq.phases}{a \code{vector} with the linkage phases between markers
in the sequence, in corresponding positions.} \item{seq.rf}{a \code{vector}
with the recombination frequencies between markers in the sequence.}
\item{seq.like}{log-likelihood of the corresponding linkage map.}
\item{data.name}{name of the object of class \code{onemap} with the raw
data.} \item{twopt}{name of the object of class \code{rf.2pts} with the
2-point analyses.}
}
\description{
Estimates the multipoint log-likelihood, linkage phases and recombination
frequencies for a sequence of markers in a given order, using sliding
windows to construct partial maps of adjacent subsets of markers
}
\details{
Markers are mapped in the order defined in the object \code{input.seq}.
A sliding window of size \code{window} runs along the sequence, in steps
of length given by argument \code{step} (which must be between \code{1}
and \code{window - 1}). For each window of markers, the best linkage phase
combination and recombination fractions are estimated and the multipoint
likelihood is calculated (for more details, check the documentation of
function \code{\link[onemap]{map}}).
}
\author{
Gabriel R A Margarido \email{gramarga@usp.br}
}
\seealso{
\code{\link[onemap]{make.seq}}. If you want to use a predefined
set of linkage phases, see \code{\link[onemap]{map}}
}
\keyword{utilities}

