\name{onemap-internal}
\alias{print.extracted.group}
\alias{print.outcross}
\alias{print.rf.3pts}
\title{Internal onemap functions}
\description{
  Currently, only some of the \code{\link{print}} functions for
  \pkg{onemap}. Not to be directly called by the user.
}
\usage{
\method{print}{extracted.group}(x, \dots)
\method{print}{outcross}(x, \dots)
\method{print}{rf.3pts}(x, \dots)
}
\author{Gabriel Rodrigues Alves Margarido}
\keyword{internal}
