% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ord_by_geno}
\alias{ord_by_geno}
\title{Order the markers in a sequence using the genomic position}
\usage{
ord_by_geno(input.seq)
}
\arguments{
\item{input.seq}{object of class `sequence`}
}
\value{
An object of class \code{sequence}
}
\description{
Order the markers in a sequence using the genomic position
}
\author{
Cristiane Taniguti
}
