% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsim2raw.R
\name{pedsim2raw}
\alias{pedsim2raw}
\title{Converts the output of PedigreeSim to onemap raw file}
\usage{
pedsim2raw(cross = c("outcross", "f2 intercross"),
  genofile = "sim_out_genotypes.dat", parent1 = "P1", parent2 = "P2",
  f1 = "F1", out.file = "sim_out.raw", miss.perc = 10)
}
\arguments{
\item{cross}{string defining the cross type "outcross" or "f2 intercross"}

\item{genofile}{pedigreeSim output .dat}

\item{parent1}{string defining the first parent ID}

\item{parent2}{string defining the seconde parent ID}

\item{f1}{string defining the F1 ID, if cross type "f2 intercross}

\item{out.file}{string defining the name of the output file}

\item{miss.perc}{double defining the percentage of missing data to be simulated}
}
\description{
Converts the output of PedigreeSim to onemap raw file
}
