% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsim2vcf.R
\name{pedsim2vcf}
\alias{pedsim2vcf}
\title{Function to simulate depths and to convert PedigreeSim file to vcf file}
\usage{
pedsim2vcf(
  inputfile = NULL,
  map.file = NULL,
  chrom.file = NULL,
  out.file = "out.pedsim2vcf.txt",
  miss.perc = 0,
  counts = TRUE,
  mean.depth = 20,
  p.mean.depth = 20,
  chr.mb = 10,
  method = c("updog", "neg.binom"),
  mean.phred = 20,
  bias = 1,
  od = 0.001,
  disper.par = 2,
  pos = NULL,
  chr = NULL,
  phase = FALSE,
  haplo.ref = NULL,
  use.as.alleles = FALSE
)
}
\arguments{
\item{inputfile}{file .dat output from PedigreeSim software}

\item{map.file}{file .map input in PedigreeSim software}

\item{chrom.file}{file.chrom input in PedigreeSim software}

\item{out.file}{path to vcf output file}

\item{miss.perc}{Percentage of missing data}

\item{counts}{If \code{TRUE} also simulates allele counts using approach defined in \code{method}}

\item{mean.depth}{mean of the negative binomial distribution to generate depth counts}

\item{p.mean.depth}{mean of the negative binomial distribution to generate depth counts for parents}

\item{chr.mb}{Chromosome size in mega base.}

\item{method}{Choose negative binomial ("neg.binom"), poisson ("poisson") distributions or updog ("updog") model to simulate counts values}

\item{mean.phred}{Sequencing error parameter}

\item{bias}{The bias parameter for updog model. Pr(a read after selected) / Pr(A read after selected).}

\item{od}{The overdispersion parameter for updog model. See the Details of the rho variable in betabinom.}

\item{disper.par}{Dispertion parameter for negative binomial}

\item{pos}{vector with the position of each marker. Use "cm" if you want to use mapfile position information.}

\item{chr}{Chromosome where the marker is positioned}

\item{phase}{if TRUE the genotypes in VCF will be phased}

\item{haplo.ref}{character indicating the reference haplotype of genotypes.dat file}

\item{use.as.alleles}{if \code{TRUE} uses codification in genotypes dat to define the reference and alternative alleles fields in VCF}
}
\value{
vcf file located in out.file defined path
}
\description{
Given PedigreeSim .dat .map and .chrom files generate vcf file with depths count
in AD format field estimated by negative binomial. The function receives genotypes codification in
genotypes.dat according with Wu et. al 2002a table I observed bands. The a allele will be considered
the reference allele. Null alleles are not supported.
}
\references{
Voorrips, R. E. , Maliepaard, C. A. (2012) The simulation of meiosis in 
diploid and tetraploid organisms using various genetic models.
\emph{BMC Bioinformatics} 13  - ISSN 1471-2105 - 12 p.

Bilton,T. P. , Schofield,  M. R. , Black,  M. A. , Chagné, D. , Wilcox, P. L. 
and Dodds, K. G. (2018) Accounting for Errors in Low Coverage High-Throughput
Sequencing Data When Constructing Genetic Maps Using Biparental Outcrossed Populations.
\emph{Genetics}. ISSN 0016-6731.
}
\seealso{
vcf file description 
<http://www.internationalgenome.org/wiki/Analysis/Variant%20Call%20Format/vcf-variant-call-format-version-40/>
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
