% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{plot.bc.onemap}
\alias{plot.bc.onemap}
\title{Draw a graphic of raw data for a backcross population}
\usage{
\method{plot}{bc.onemap}(x)
}
\arguments{
\item{x}{an object of class bc.onemap, with data and additional information}
}
\value{
a ggplot graphic
}
\description{
Shows a heatmap like graphic (in ggplot2, a "tile") for raw data.
Lines correspond to markers and columns for individuals.
The function receives a onemap object of class bc.onemap, reads information
from genotypes from this object, convert it to a long dataframe format
using function melt() from package reshape2(), converts numbers from the object
to genetic notation (AA, AB, -), then plot the graphic.
If there is more than 20 markers, removes y labels
}
\examples{
data(fake.bc.onemap) # Loads a fake backcross dataset installed with onemap
plot(fake.bc.onemap) # This will show you the graph

# You can store the graphic in an object, then save it.
# For details, see the help of ggplot2's function ggsave()
g <- plot(fake.bc.onemap)
ggsave("MyRawData.jpg", g, width=7, height=4, dpi=600)
}

