% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.raw.data.R
\name{plot.by.type}
\alias{plot.by.type}
\title{plot.by.type}
\usage{
\method{plot}{by.type}(x, subcateg = TRUE)
}
\arguments{
\item{x}{an object of class outcross, f2.onemap, bc.onemap, risib.onemap or riself.onemap}

\item{subcateg}{a TRUE/FALSE option to indicate if results will be plotted showing
all possible categories (only for outcrossing populations)}
}
\value{
a ggplot graphic
}
\description{
Draw a graphic showing the number of markers on each segregation pattern.
The function receives a onemap object of class outcross, f2.onemap, bc.onemap,
risib.onemap or riself.onemap.
For outcrossing populations, it can show detailed information (all 18 possible categories)
or a simplified version.
}
\examples{
data(example.out) #Outcrossing data
plot.by.type(example.out)
plot.by.type(example.out, subcateg=FALSE)

data(fake.bc.onemap)
plot.by.type(fake.bc.onemap)

data(fake.f2.onemap)
plot.by.type(fake.f2.onemap)

# You can store the graphic in an object, then save it.
# For details, see the help of ggplot2's function ggsave()
data(example.out) #Outcrossing data
g <- plot.by.type(example.out)
ggsave("SegregationTypes.jpg", g, width=7, height=4, dpi=600)
}

