% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_haplotypes.R
\name{plot.onemap_progeny_haplotypes}
\alias{plot.onemap_progeny_haplotypes}
\title{Plots progeny haplotypes}
\usage{
\method{plot}{onemap_progeny_haplotypes}(
  probs,
  col = NULL,
  position = "stack",
  show_markers = TRUE,
  main = "Genotypes"
)
}
\arguments{
\item{probs}{object of class onemap_progeny_haplotypes}

\item{col}{Color of parentes' homologous.}

\item{position}{"split" or "stack"; if "split" (default) the parents' homologous are plotted separately. if "stack" the parents' homologous are plotted together.}

\item{show_markers}{logical; if  \code{TRUE}, the markers (default) are plotted.}

\item{main}{An overall title for the plot; default is \code{NULL}.}
}
\description{
Plots progeny haplotypes
}
\author{
Getulio Caixeta Ferreira, \email{getulio.caifer@gmail.com}

Cristiane Taniguti, \email{chtaniguti@usp.br}
}
