% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.raw.data.R
\name{plot.outcross}
\alias{plot.outcross}
\title{plot.outcross}
\usage{
\method{plot}{outcross}(x, all = TRUE)
}
\arguments{
\item{x}{an object of class outcross, with data and additional information}

\item{all}{a TRUE/FALSE option to indicate if results will be plotted together (if TRUE)
or splitted based on their segregation pattern}
}
\value{
a ggplot graphic
}
\description{
Draw a graphic of raw data for an outcross  population, using ggplot2.
Lines correspond to markers and columns for individuals.
The function can plot a graph for all types of markers.
The graphic is a "heatmap", whose name in ggplot2 is "tile".
The function receives a onemap object of class outcross, reads information
from genotypes from this object, convert it to a long dataframe format
using onemap internal function create.dataframe.outcross.plot(),
then plot the graphic.
If there is more than 20 markers, removes y labels
It can show all markers together, or it can split them according the segregation
pattern.
}
\examples{
data(example.out) # Loads a fake backcross dataset installed with onemap
plot(example.out) # This will show you the graph for all markers
plot(example.out, all=FALSE) # This will show you the graph splitted for marker types

# You can store the graphic in an object, then save it.
# For details, see the help of ggplot2's function ggsave()
g <- plot(example.out, all=FALSE)
ggsave("MyRawData.jpg", g, width=9, height=4, dpi=600)
}

