% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{plot.riself.onemap}
\alias{plot.riself.onemap}
\title{Draw a graphic of raw data for a RIL population (made by selfing)}
\usage{
\method{plot}{riself.onemap}(x)
}
\arguments{
\item{x}{an object of class riself.onemap, with data and additional information}
}
\value{
a ggplot graphic
}
\description{
Draw a graphic of raw data for a RIL population (made by selfing), using ggplot2.
Lines correspond to markers and columns for individuals.
The graphic is a "heatmap", whose name in ggplot2 is "tile".
The function receives a onemap object of class riself.onemap, reads information
from genotypes from this object, convert it to a long dataframe format
using function melt() from package reshape2(), converts numbers from the object
to genetic notation (AA, BB, -), then plot the graphic.
If there is more than 20 markers, removes y labels
}
\examples{
# Please, see examples for plot.bc.onemap; the only difference is the dataset
}

