% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.raw.data.R
\name{plot.risib.onemap}
\alias{plot.risib.onemap}
\title{plot.risib.onemap}
\usage{
\method{plot}{risib.onemap}(x)
}
\arguments{
\item{x}{an object of class risib.onemap, with data and additional information}
}
\value{
a ggplot graphic
}
\description{
Draw a graphic of raw data for a RIL population (made by sibing), using ggplot2.
In fact, the graphic for the raw data will have the same aspect of the ones
for RILs made by selfing. Therefore, this function will only call
plot.riself.onemap()
}
\examples{
# Please, see examples for plot.bc.onemap; the only difference is the dataset
}

