% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_error.R
\name{plot_error_dist}
\alias{plot_error_dist}
\title{Plot a barplot with error probabilities values}
\usage{
plot_error_dist(onemap.obj = NULL, mk.type = TRUE, select.ind = NULL,
  select.mk = NULL, n.int = NULL)
}
\arguments{
\item{onemap.obj}{an object of class \code{onemap} coming from \code{read_onemap}, 
\code{read_mapmaker}, \code{onemap_read_vcfR}, \code{updog_error} functions}

\item{mk.type}{a TRUE/FALSE value to define if genotypes will colored by marker type}

\item{select.ind}{a string defining specific individuals. The graphic will only contains 
error probability information of this individuals.}

\item{select.mk}{a string defining specific markers. The graphic will only contains 
error probability information of this markers.}

\item{n.int}{}
}
\description{
Plot a barplot with error probabilities values
}
\examples{

\dontrun{
data(onemap_example_out)
p <- plot_error_dist(onemap_example_out)
ggplot2::ggsave("errors_out.jpg", p width=7, height=4, dpi=300)
}
}
