% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyRAD_error.R
\name{polyRAD_error}
\alias{polyRAD_error}
\title{OneMap interface with polyRAD package}
\usage{
polyRAD_error(vcf = NULL, onemap.obj = NULL, parent1 = NULL,
  parent2 = NULL, f1 = NULL, crosstype = NULL, tech.issue = TRUE)
}
\arguments{
\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}
}
\value{
onemap object with error updated
}
\description{
Put here a description
}
\references{
Clark LV, Lipka AE, and Sacks EJ (2019) Improvements to 
Genotype Calling in Polyploids Using the polyRAD R Package. 
Plant and Animal Genome Conference XXVII, January 12-16, 
San Diego, California, USA. doi:10.13140/RG.2.2.18358.75847

Clark LV, Lipka AE, and Sacks EJ (2018) polyRAD: Genotype 
Calling with Uncertainty from Sequencing Data in Polyploids 
and Diploids. Plant and Animal Genome Conference XXVI, 
January 13-17, San Diego, California, USA. doi:10.13140/RG.2.2.27134.08001
}
\seealso{
\code{\link[onemap]{extract_depth}} 
    \code{\link[onemap]{binom_error}} and 
    \url{https://github.com/dcgerard/updog}.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
