% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyRAD_genotype.R
\name{polyRAD_genotype}
\alias{polyRAD_genotype}
\title{OneMap interface with polyRAD package}
\usage{
polyRAD_genotype(
  vcf = NULL,
  onemap.obj = NULL,
  parent1 = NULL,
  parent2 = NULL,
  f1 = NULL,
  crosstype = NULL,
  global_error = NULL,
  use_genotypes_errors = TRUE,
  use_genotypes_probs = FALSE
)
}
\arguments{
\item{vcf}{path to vcf file}

\item{onemap.obj}{object of onemap class}

\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{f1}{f1 individual identification if F2 cross type}

\item{crosstype}{string defining the cross type, by now it supports only 
outcross and f2 intercross}

\item{global_error}{number from 0 to 1 defining the global error to be considered together 
with the genotype errors or the genotype probabilities or NULL to not considered any global error}

\item{use_genotypes_errors}{if \code{TRUE} the error probability of each genotype will be considered in emission function of HMM}

\item{use_genotype_probs}{if \code{TRUE} the probability of each possible genotype will be considered in emission function of HMM}
}
\value{
onemap object with error updated
}
\description{
Put here a description
}
\references{
Clark LV, Lipka AE, and Sacks EJ (2019) Improvements to 
Genotype Calling in Polyploids Using the polyRAD R Package. 
Plant and Animal Genome Conference XXVII, January 12-16, 
San Diego, California, USA. doi:10.13140/RG.2.2.18358.75847

Clark LV, Lipka AE, and Sacks EJ (2018) polyRAD: Genotype 
Calling with Uncertainty from Sequencing Data in Polyploids 
and Diploids. Plant and Animal Genome Conference XXVI, 
January 13-17, San Diego, California, USA. doi:10.13140/RG.2.2.27134.08001
}
\seealso{
\code{\link[onemap]{extract_depth}} 
    \code{\link[onemap]{binom_error}} and 
    \url{https://github.com/dcgerard/updog}.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
