% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.compare}
\alias{print.compare}
\title{print method for object class 'compare'}
\usage{
\method{print}{compare}(x, ...)
}
\arguments{
\item{x}{object of class compare}

\item{...}{currently ignored}
}
\value{
compare object description
}
\description{
print method for object class 'compare'
}
