% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group.R
\name{print.group}
\alias{print.group}
\title{Show the results of grouping procedure}
\usage{
\method{print}{group}(x, detailed = TRUE, ...)
}
\arguments{
\item{x}{an object of class onemap.segreg.test}

\item{detailed}{logical. If \code{TRUE} the markers in each
    linkage group are printed.}

\item{...}{currently ignored}
}
\value{
\code{NULL}
}
\description{
It shows the linkage groups as well as the unlinked markers.
}
\keyword{internal}

