% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_upgma.R
\name{print.group.upgma}
\alias{print.group.upgma}
\title{Show the results of grouping procedure}
\usage{
\method{print}{group.upgma}(x, ...)
}
\arguments{
\item{x}{an object of class group.upgma}

\item{...}{currently ignored}
}
\value{
\code{NULL}
}
\description{
It shows the linkage groups as well as the unlinked markers.
}
\keyword{internal}
