% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_seq.R
\name{print.group_seq}
\alias{print.group_seq}
\title{Show the results of grouping markers to preexisting sequence}
\usage{
\method{print}{group_seq}(x, detailed = TRUE)
}
\arguments{
\item{x}{an object of class group_seq}

\item{detailed}{logical. If \code{TRUE} the markers in each
linkage group sequence are printed.}
}
\value{
\code{NULL}
}
\description{
It shows the groups sequences, the repeated markers, as well as the unlinked markers.
}
\keyword{internal}
