% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_onemap.R
\name{print.onemap}
\alias{print.onemap}
\title{Print method for object class 'onemap'}
\usage{
\method{print}{onemap}(x, ...)
}
\arguments{
\item{x}{object of class onemap}

\item{...}{currently ignored}
}
\value{
printed information about onemap object
}
\description{
Print method for object class 'onemap'
}
