% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bins.R
\name{print.onemap_bin}
\alias{print.onemap_bin}
\title{print method for object class 'onemap_bin'}
\usage{
\method{print}{onemap_bin}(x, ...)
}
\arguments{
\item{x}{object of class \code{onemap_bin}}

\item{...}{currently ignored}
}
\value{
No return value, called for side effects
}
\description{
print method for object class 'onemap_bin'
}
