% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_segregation.R
\name{print.onemap_segreg_test}
\alias{print.onemap_segreg_test}
\title{Show the results of segregation tests}
\usage{
\method{print}{onemap_segreg_test}(x, ...)
}
\arguments{
\item{x}{an object of class onemap_segreg_test}

\item{...}{currently ignored}
}
\value{
a dataframe with marker name, H0 hypothesis, chi-square statistics,
p-values, and % of individuals genotyped.
}
\description{
It shows the results of Chisquare tests performed for all markers in a onemap object
of cross type outcross, backcross, F2 intercross or recombinant inbred lines.
}
\examples{

data(onemap_example_out) # Loads a fake outcross dataset installed with onemap
Chi <- test_segregation(onemap_example_out) # Performs the chi-square test for all markers
print(Chi) # Shows the results

}
