% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_seq.R
\name{print.order}
\alias{print.order}
\title{Print order_seq object}
\usage{
\method{print}{order}(x, ...)
}
\arguments{
\item{x}{object of class order_seq}

\item{...}{currently ignored}
}
\value{
printed information about order_seq object
}
\description{
Print order_seq object
}
