% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rf.2pts.R
\name{print.rf.2pts}
\alias{print.rf.2pts}
\title{Print method for object class 'rf.2pts'}
\usage{
\method{print}{rf.2pts}(x, mrk = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{rf.2pts}.}

\item{mrk}{a vector containing a pair of markers, so detailed
    results of the two-point analysis will be printed for them.
    Can be numeric or character strings indicating the
    numbers/names corresponding to any markers in the input file.}

\item{...}{further arguments, passed to other methods. Currently ignored.}
}
\value{
\code{NULL}
}
\description{
It shows the linkage groups as well as the unlinked markers.
}
\keyword{internal}

