% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_2pts.R
\name{print.rf_2pts}
\alias{print.rf_2pts}
\title{Print method for object class 'rf_2pts'}
\usage{
\method{print}{rf_2pts}(x, mrk = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{rf_2pts}.}

\item{mrk}{a vector containing a pair of markers, so detailed
results of the two-point analysis will be printed for them.
Can be numeric or character strings indicating the
numbers/names corresponding to any markers in the input file.}

\item{...}{further arguments, passed to other methods. Currently ignored.}
}
\value{
\code{NULL}
}
\description{
It shows the linkage groups as well as the unlinked markers.
}
\keyword{internal}

