% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seq.R
\name{print.sequence}
\alias{print.sequence}
\title{Print method for object class 'sequence'}
\usage{
\method{print}{sequence}(x, ...)
}
\arguments{
\item{x}{object of class sequence}

\item{...}{corrently ignored}
}
\value{
printed information about sequence object
}
\description{
Print method for object class 'sequence'
}
