% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/try_seq.R
\name{print.try}
\alias{print.try}
\title{Print method for object class 'try'}
\usage{
\method{print}{try}(x, j = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{try}.}

\item{j}{if \code{NULL} (default), output is a summary of the
results for all possible positions of the additional
marker. Otherwise, an integer makes detailed output to be
printed for the corresponding position. This integer must be
less than or equal to the length of the original sequence plus
1.}

\item{...}{currently ignored}

\item{\dots}{further arguments, passed to other
methods. Currently ignored.}
}
\value{
\code{NULL}
}
\description{
Print method for object class 'try'
}
\keyword{internal}

