% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_haplotypes.R
\name{progeny_haplotypes_wide_format}
\alias{progeny_haplotypes_wide_format}
\title{Wide table format for progeny haplotypes}
\usage{
progeny_haplotypes_wide_format(input_seq, allele = c("ab", "SNP"))
}
\arguments{
\item{input_seq}{object of class `sequence`}

\item{character}{defining the haplotype alleles codification. 
If "SNP" the ab code format is replaced by the SNP base, 
if the information is available in the input object.}
}
\value{
data.frame with marker, group, map position and haplotype alleles for each individuals in the progeny
}
\description{
Wide table format for progeny haplotypes
}
