% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mapmaker.R
\name{read.mapmaker}
\alias{read.mapmaker}
\title{Read data from a Mapmaker raw file}
\usage{
read.mapmaker(dir, file)
}
\arguments{
\item{dir}{directory where the input file is located.}

\item{file}{the name of the input file which contains the data to be read.}
}
\value{
An object of class \code{onemap}, i.e., a list with the following
components: \item{geno}{a matrix with integers indicating the genotypes
read for each marker in \code{onemap} fashion. Each column contains data
for a marker and each row represents an individual.}

\item{geno.mmk}{ a matrix with
integers indicating the genotypes read for each marker in
\code{MAPMAKER/EXP} fashion, i.e., 1, 2, 3: AA, AB, BB, respectively; 3, 4:
BB, not BB, respectively; 1, 5: AA, not AA, respectively. Each column
contains data for a marker and each row represents an individual.}

\item{n.ind}{number of individuals.} \item{n.mar}{number of markers.}
\item{segr.type}{a vector with the segregation type of each marker, as
\code{strings}. Segregation types were adapted from outcross segregation
types, using the same notation. For details see \link{read.onemap}.}
\item{segr.type.num}{a vector with the segregation type of each marker,
represented in a simplified manner as integers. Segregation types were
adapted from outcross segregation types. For details see
\link{read.onemap}.} \item{input}{the name of the input file.}
\item{n.phe}{number of phenotypes.} \item{pheno}{a matrix with phenotypic
values.  Each column contains data for a trait and each row represents an
individual. Currently ignored.}
}
\description{
Imports data from a Mapmaker raw file.
}
\details{
For details about MAPMAKER files see \cite{Lincoln et al.} (1993). The
current version supports backcross, F2s and RIL populations. The file
can contain phenotypic data, but it will not be used in the analysis.
}
\examples{

  \dontrun{
    map_data <-read.mapmaker(dir="work_directory",file="data_file.txt")
    #Checking 'fake.f2.onemap'
    data(fake.f2.onemap)
    names(fake.f2.onemap)
  }

}
\author{
Adapted from Karl Broman (package \pkg{qtl}) by Marcelo Mollinari,
\email{mmollina@usp.br}
}
\references{
Broman, K. W., Wu, H., Churchill, G., Sen, S., Yandell, B.
(2008) \emph{qtl: Tools for analyzing QTL experiments} R package version
1.09-43

Lincoln, S. E., Daly, M. J. and Lander, E. S. (1993) Constructing genetic
linkage maps with MAPMAKER/EXP Version 3.0: a tutorial and reference
manual. \emph{A Whitehead Institute for Biomedical Research Technical
Report}.
}
\seealso{
\code{fake.bc.onemap} and \code{fake.f2.onemap} directory in the
package source.
}
\keyword{IO}

