% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_inds}
\alias{remove_inds}
\title{Remove inviduals from the onemap object}
\usage{
remove_inds(onemap.obj, rm.ind)
}
\arguments{
\item{onemap.obj}{object of class onemap}

\item{rm.ind}{vector of charaters with individuals names}
}
\description{
Remove inviduals from the onemap object
}
