% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_dupli_mks}
\alias{rm_dupli_mks}
\title{Remove duplicated markers keeping the one with less missing data
 
 @param onemap.obj object of class \code{onemap}
 
 @export}
\usage{
rm_dupli_mks(onemap.obj)
}
\description{
Remove duplicated markers keeping the one with less missing data
 
 @param onemap.obj object of class \code{onemap}
 
 @export
}
