% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pedsim.R
\name{run_pedsim}
\alias{run_pedsim}
\title{Generate input files for PedigreeSim software}
\usage{
run_pedsim(
  chromosome = c("Chr1", "Chr2", "Chr3"),
  n.marker = c(18, 19, 20),
  tot.size.cm = c(18, 19, 20),
  centromere = c(9, 9, 10),
  n.ind = 150,
  mk.types = c("A1", "D1.10", "D2.15"),
  n.types = c(7, 25, 25),
  pop = c("F1", "F2", "BC"),
  path.pedsim = NULL,
  name.mapfile = "mapfile.map",
  name.founderfile = "founderfile.gen",
  name.chromfile = "sim.chrom",
  name.parfile = "sim.par",
  name.out = "sim_out",
  rm.tempfiles = F
)
}
\arguments{
\item{chromosome}{a vector of characters defining the chromosomes names}

\item{n.marker}{a vector of integers defining the number of markers within chromosome}

\item{tot.size.cm}{a vector of double numbers defining the total size in centimorgan of each chromosome}

\item{centromere}{a vector of double numbers defining the centromere position}

\item{n.ind}{a integer defining the number of individuals to be simulated in the progeny}

\item{mk.types}{a vector of characters defining the marker types according with Wu2002a 
(the types will be distribuited randomly by the chromosomes)}

\item{n.types}{a vector of integers defining the number of markers of each type}

\item{pop}{a character defining the population type. Supported: "F1", "F2", "BC"}

\item{path.pedsim}{path to the .jar PedigreeSim file}

\item{name.mapfile}{a character with the mapfile name}

\item{name.founderfile}{a character with the founderfile name}

\item{name.chromfile}{a character with the chromfile name}

\item{name.parfile}{a character with the parfile name}

\item{name.out}{a character with the PedigreeSim outputs names}

\item{rm.tempfiles}{logical TRUE/FALSE, if TRUE remove the temporary files generated}
}
\value{
PedigreeSim outputs including the population genotypes
}
\description{
With the function arguments users can define the basic characteristics 
of the maps to be simulated. The PedigreeSim software must be located at 
path.pedsim. The PedigreeSim software offers other options that are not available here.
}
\references{
Wu, R., Ma, C.-X., Painter, I. and Zeng, Z.-B. (2002)
Simultaneous maximum likelihood estimation of linkage and linkage phases in
outcrossing species. \emph{Theoretical Population Biology} 61: 349-363.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
