% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{save_onemap_sequences}
\alias{save_onemap_sequences}
\title{Save a list of onemap sequence objects}
\usage{
save_onemap_sequences(sequences.list, filename)
}
\arguments{
\item{sequences.list}{list of \code{sequence} objects}

\item{filename}{name of the output file (Ex: my_beautiful_map.RData)}
}
\description{
The onemap sequence object contains everything users need to reproduce the complete analysis:
the input onemap object, the rf_2pts result, and the sequence genetic distance and marker order.
Therefore, a list of sequences is the only object users need to save to be able to recover all analysis.
But simple saving the list of sequences will save many redundant objects. This redundancy is only considered by R
when saving the object. For example, one input object and the rf_2pts result will be saved for every sequence.
}
