% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_by_type}
\alias{seq_by_type}
\title{Extract marker number by name}
\usage{
seq_by_type(x, marker.type)
}
\arguments{
\item{x}{object of class or sequence}

\item{marker.type}{vector of character with marker type to be selected}
}
\value{
new sequence object with selected marker type
}
\description{
Extract marker number by name
}
