% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_func.R
\name{set_map_fun}
\alias{set_map_fun}
\title{Defines the default mapping function}
\usage{
set_map_fun(type = c("kosambi", "haldane"))
}
\arguments{
\item{type}{Indicates the function that should be used, which can be
\code{"kosambi"} or \code{"haldane"}}
}
\description{
Defines the function that should be used to display the genetic map through
the analysis.
}
\author{
Marcelo Mollinari, \email{mmollina@usp.br}
}
\references{
Haldane, J. B. S. (1919) The combination of linkage values and
the calculation of distance between the loci of linked factors.
\emph{Journal of Genetics} 8: 299-309.

Kosambi, D. D. (1944) The estimation of map distance from recombination
values. \emph{Annuaire of Eugenetics} 12: 172-175.
}
\seealso{
\code{\link[onemap]{kosambi}} and \code{\link[onemap]{haldane}}
}
\keyword{arith}

