% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_by_pos}
\alias{sort_by_pos}
\title{Sort markers in onemap object by their position in reference genome}
\usage{
sort_by_pos(onemap.obj)
}
\arguments{
\item{onemap.obj}{object of class onemap}
}
\description{
Sort markers in onemap object by their position in reference genome
}
