% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_2pts}
\alias{split_2pts}
\title{Split rf_2pts object by mks}
\usage{
split_2pts(twopts.obj, mks)
}
\arguments{
\item{twopts.obj}{object of class rf_2pts}

\item{mks}{markers names (vector of characters) or number (vector of integers) to be removed and added to a new rf_2pts object}
}
\description{
Split rf_2pts object by mks
}
