% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_onemap.R
\name{split_onemap}
\alias{split_onemap}
\title{Split onemap data sets}
\usage{
split_onemap(onemap.obj = NULL, mks = NULL)
}
\arguments{
\item{onemap.obj}{object of class onemap}

\item{mks}{markers names (vector of characters) or number (vector of integers) to be removed and added to a new onemap object}
}
\value{
a list containing in first level the original onemap object without 
the indicated markers and the second level the new onemap object with only 
the indicated markers
}
\description{
Receives one onemap object and a vector with markers names to be 
removed from the input onemap object and inserted in a new one. The output
is a list containing the two onemap objects.
}
