% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_maps.R
\name{summary_maps_onemap}
\alias{summary_maps_onemap}
\title{Create table with summary information about the linkage map}
\usage{
summary_maps_onemap(map.list, mapping_function = "kosambi")
}
\arguments{
\item{map.list}{a map, i.e. an object of class \code{sequence} with a
predefined order, linkage phases, recombination fraction and likelihood;
also it could be a list of maps.}

\item{mapping_function}{either "kosambi" or "haldane"}
}
\value{
data.frame with basic summary statistics
}
\description{
Create table with summary information about the linkage map
}
\author{
Jeekin Lau, \email{jeekinlau@gmail.com}
}
