% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updog_error.R
\name{updog_error}
\alias{updog_error}
\title{OneMap interface with updog package}
\usage{
updog_error(vcfR.object = NULL, onemap.object = NULL,
  vcf.par = c("AD", "DPR"), parent1 = "P1", parent2 = "P2",
  f1 = NULL, recovering = FALSE, mean_phred = 20, cores = 2,
  depths = NULL)
}
\arguments{
\item{parent1}{parent 1 identification in vcfR object}

\item{parent2}{parent 2 identification in vcfR objetc}

\item{depths}{list containing a matrix for ref and other for alt allele counts, samples ID in colum and markers ID in rows}
}
\value{
onemap object with error updated
}
\description{
Uses alelle counts to re-estimate genotypes with updog approach and 
considers the genotypes probabilities in for further multipoint 
analysis
}
\references{
Gerard, D., Ferrão L.F.V., Garcia, A.A.F., & Stephens, M. (2018). Harnessing 
Empirical Bayes and Mendelian Segregation for Genotyping Autopolyploids from 
Messy Sequencing Data. bioRxiv. doi: 10.1101/281550.
}
\seealso{
\code{\link[onemap]{extract_depth}} 
    \code{\link[onemap]{binom_error}} and 
    \url{https://github.com/dcgerard/updog}.
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
