% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_haplotypes.R
\name{vcf2progeny_haplotypes}
\alias{vcf2progeny_haplotypes}
\title{Convert phased vcf to onemap_progeny_haplotypes object. Turns possible to use
the plot.onemap_progeny_haplotypes (By now only for outcrossing population).}
\usage{
vcf2progeny_haplotypes(
  vcfR.object,
  ind.id = NULL,
  group_names = NULL,
  parent1,
  parent2,
  crosstype
)
}
\arguments{
\item{vcfR.object}{vcfR object}

\item{ind.id}{vector of characters with progeny individuals to be evaluated}

\item{group_names}{string with chromosomes or group names to be evaluated}

\item{parent1}{character with parent 1 ID}

\item{parent2}{character with parent 2 ID}

\item{crosstype}{character defining crosstype (outcross, f2 intercross, f2 backcross, ril sib, ril self)}
}
\description{
Convert phased vcf to onemap_progeny_haplotypes object. Turns possible to use
the plot.onemap_progeny_haplotypes (By now only for outcrossing population).
}
