% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcf2raw.R
\name{vcf2raw}
\alias{vcf2raw}
\title{Convert variants from a VCF file to OneMap file format}
\usage{
vcf2raw(input = NULL, output = NULL, parent1 = NULL, parent2 = NULL,
  min_class = 1)
}
\arguments{
\item{input}{path to the input VCF file.}

\item{output}{path to the output OneMap file.}

\item{parent1}{\code{string} or \code{vector} of \code{strings} specifying
sample ID(s) of the first progeny parent.}

\item{parent2}{\code{string} or \code{vector} of \code{strings} specifying
sample ID(s) of the second progeny parent.}

\item{min_class}{a real number between 0.0 and 1.0. For each parent and each
variant site, at least \eqn{parent sample count \times min_class} of the parent
samples must be of the same genotype for it to be assigned to the corresponding
parent.}
}
\description{
Converts data from a standard VCF (Variant Call Format) file to the input
format required by OneMap, while trying to identify the appropriate
segregation pattern.
}
\details{
Each variant in the VCF file is processed independently. Only biallelic SNPs
and indels for diploid variant sites are considered.

First, samples corresponding to both parents of the progeny are parsed and
their genotypes identified, given that they are concordant above a provided
threshold. Marker type (B3, D1 or D2) is determined based on parental genotypes.
Finally, progeny genotypes are identified and output is produced. Variants
for which parent genotypes cannot be determined are discarded.

Reference sequence ID and position for each variant site are stored as special
phenotypes denoted \code{CHROM} and \code{POS}.
}
\examples{
\dontrun{
    vcf2raw(input="your_VCF_file.vcf", output="your_OneMap_file.raw",
            parent1=c("PAR1_sample1", "PAR1_sample2"),
            parent2=c("PAR2_sample1", "PAR2_sample2", "PAR2_sample3"),
            min_class=0.5)
  }
}
\author{
Gabriel R A Margarido, \email{gramarga@gmail.com}
}
\seealso{
\code{read.outcross} for a description of the OneMap file format.
}
\keyword{IO}

