% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_bins.R
\name{add_redundants}
\alias{add_redundants}
\title{Add the redundant markers removed by create_data_bins function}
\usage{
add_redundants(sequence, onemap.obj, bins)
}
\arguments{
\item{sequence}{object of class \code{sequence}}

\item{onemap.obj}{object of class \code{onemap.obj} before redundant markers were removed}

\item{bins}{object of class \code{onemap_bin}}
}
\description{
Add the redundant markers removed by create_data_bins function
}
