% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_probs.R
\name{create_probs}
\alias{create_probs}
\title{Converts probabilities to onemap format}
\usage{
create_probs(
  onemap.obj = NULL,
  global_error = NULL,
  genotypes_errors = NULL,
  genotypes_probs = NULL
)
}
\arguments{
\item{onemap.obj}{define}

\item{global_error}{define}

\item{genotypes_errors}{define}

\item{genotypes_probs}{define}
}
\description{
Converts probabilities to onemap format
}
