% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_2pts_gaps}
\alias{filter_2pts_gaps}
\title{Filter markers based on 2pts distance}
\usage{
filter_2pts_gaps(input.seq, max.gap = 10)
}
\arguments{
\item{input.seq}{object of class sequence with ordered markers}

\item{max.gap}{maximum gap measured in kosambi centimorgans allowed between adjacent markers. 
Markers that presents the defined distance between both adjacent neighbors will be removed.}
}
\description{
Filter markers based on 2pts distance
}
