% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bins_by_prob.R
\name{find_bins_by_probs}
\alias{find_bins_by_probs}
\title{Not ready yet - use carefully}
\usage{
find_bins_by_probs(onemap_obj, threshold.probs = 1e-04, threshold.count = 0.08)
}
\description{
Not ready yet - use carefully
}
