% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping_batches.R
\name{generate_overlapping_batches}
\alias{generate_overlapping_batches}
\title{Function to divide the sequence in batches with user defined size}
\usage{
generate_overlapping_batches(
  input.seq,
  size = 50,
  overlap = 15,
  silent = FALSE
)
}
\description{
Function to divide the sequence in batches with user defined size
}
