% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_seq.R
\name{group_seq}
\alias{group_seq}
\title{Function to assign markers to preexisting linkage groups
Identifies linkage groups of markers, combining input \code{sequences} objects with
markers other markers from \code{rf_2pts} object. The results from two-point
(pairwise) analysis and the \emph{transitive} property of linkage are used for
grouping, as \code{group} function.}
\usage{
group_seq(input.2pts, seqs = "CHROM", unlink.mks = "all",
  rm.repeated = FALSE, LOD = NULL, max.rf = NULL)
}
\arguments{
\item{input.2pts}{an object of class \code{rf_2pts}.}

\item{seqs}{a list of objects of class \code{sequence} or the string
"CHROM" if there is \code{CHROM} information available in the input
data file.}

\item{unlink.mks}{a object of class \code{sequence} with the number of
the markers to be grouped with the preexisting sequences defined by \code{seqs}
parameter. Using the string "all", all remaining markers of
the \code{rf_2pts} object will be tested.}

\item{rm.repeated}{logical. If \code{TRUE}, markers grouped in more than
one of the sequences are keeped in the output sequences. If \code{FALSE},
they are removed of the output sequences.}

\item{LOD}{a (positive) real number used as minimum LOD score
(threshold) to declare linkage.}

\item{max.rf}{a real number (usually smaller than 0.5) used as
maximum recombination fraction to declare linkage.}
}
\value{
A list containing the following components:\item{sequences}{a
list of objects of class \code{sequence}}. \item{repeated}{a list with the
repeated markers numbers by sequence.}
}
\description{
If the arguments specifying thresholds used to group markers, i.e., minimum
LOD Score and maximum recombination fraction, are \code{NULL} (default),
the values used are those contained in object \code{input.2pts}. If not
using \code{NULL}, the new values override the ones in object
\code{input.2pts}.
}
\examples{
data(example_out) # load OneMap's fake dataset for a outcrossing population
data(vcf_example_out) # load OneMap's fake dataset from a VCF file for a outcrossing population
comb_example <- combine_onemap(example_out, vcf_example_out) # Combine datasets
twopts <- rf_2pts(comb_example)

out_CHROM <- group_seq(twopts, seqs="CHROM", rm.repeated=FALSE)
out_CHROM

seq1 <- make_seq(twopts, c(1,2,3,4,5,25,26))
seq2 <- make_seq(twopts, c(8,18))
seq3 <- make_seq(twopts, c(4,16,20,21,24,29))

out_seqs <- group_seq(twopts, seqs=list(seq1,seq2,seq3))
out_seqs


}
\seealso{
\code{\link[onemap]{make_seq}} and \code{\link[onemap]{group}}
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
