% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_func.R
\name{haldane}
\alias{haldane}
\title{Apply haldane function to recombination fraction values}
\usage{
haldane(rcmb)
}
\description{
@param rcmb vector with recombination fraction values
 
 @return vector with centimorgan values
}
