% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_func.R
\name{kosambi}
\alias{kosambi}
\title{Apply Kosambi mapping function}
\usage{
kosambi(rcmb)
}
\arguments{
\item{rcmb}{vector of recombination fraction values}
}
\value{
vector with centimorgan values
}
\description{
Apply Kosambi mapping function
}
