% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_avoid_unlinked}
\alias{map_avoid_unlinked}
\title{Repeat HMM if map find unlinked marker}
\usage{
map_avoid_unlinked(
  input.seq,
  size = NULL,
  overlap = NULL,
  phase_cores = 1,
  tol = 1e-04,
  parallelization.type = "PSOCK",
  max.gap = FALSE,
  global_error = NULL,
  genotypes_errors = NULL,
  genotypes_probs = NULL
)
}
\arguments{
\item{input.seq}{object of class sequence}

\item{size}{The center size around which an optimum is to be searched}

\item{overlap}{The desired overlap between batches}

\item{phase_cores}{The number of parallel processes to use when estimating
the phase of a marker. (Should be no more than 4)}

\item{tol}{tolerance for the C routine, i.e., the value used to evaluate
convergence.}

\item{parallelization.type}{one of the supported cluster types. This should 
be either PSOCK (default) or FORK.}

\item{max.gap}{the marker will be removed if it have gaps higher than this defined threshold in both sides}

\item{global_error}{single value to be considered as error probability in HMM emission function}

\item{genotypes_errors}{matrix individuals x markers with error values for each marker}

\item{genotypes_probs}{table containing the probability distribution for each combination of marker × individual. 
Each line on this table represents the combination of one marker with one individual, and the respective probabilities.
The table should contain four three columns (prob(AA), prob(AB) and prob(BB)) and individuals*markers rows.}
}
\description{
Repeat HMM if map find unlinked marker
}
