% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_avoid_unlinked}
\alias{map_avoid_unlinked}
\title{Repeat HMM if map find unlinked maker}
\usage{
map_avoid_unlinked(
  input.seq,
  size = NULL,
  overlap = NULL,
  round = 5,
  phase_cores = 1,
  tol = 1e-05
)
}
\arguments{
\item{input.seq}{object of class sequence}

\item{size}{The center size around which an optimum is to be searched}

\item{overlap}{The desired overlap between batches}

\item{phase_cores}{The number of parallel processes to use when estimating
the phase of a marker. (Should be no more than 4)}

\item{tol}{tolerance for the C routine, i.e., the value used to evaluate
convergence.}
}
\description{
Repeat HMM if map find unlinked maker
}
