% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_seq.R
\name{parents_haplotypes}
\alias{parents_haplotypes}
\title{Export parents estimated haplotypes to a text file}
\usage{
parents_haplotypes(x, out_file = "test.txt")
}
\arguments{
\item{x}{sequence or list of sequences}

\item{out_file}{character defining the output file name}
}
\description{
Export parents estimated haplotypes to a text file
}
