% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedsim2vcf.R
\name{pedsim2vcf}
\alias{pedsim2vcf}
\title{Function to simulate depths and to convert PedigreeSim file to vcf file}
\usage{
pedsim2vcf(inputfile = NULL, map.file = NULL, chrom.file = NULL,
  out.file = "out.pedsim2vcf.txt", miss.perc = 0, counts = TRUE,
  mean.depth = 20, p.mean.depth = 20, disper.par = 2, chr.mb = 10,
  method = c("updog", "neg.binom", "poisson"), mean.phred = 20,
  bias = 0, od = 0, pos = NULL, haplo.ref = NULL, chr = NULL,
  phase = FALSE)
}
\arguments{
\item{inputfile}{file .dat output from PedigreeSim software}

\item{map.file}{file .map input in PedigreeSim software}

\item{chrom.file}{file.chrom input in PedigreeSIm software}

\item{out.file}{path to vcf output file}

\item{miss.perc}{Percentage of missing data}

\item{mean.depth}{mean of the negative binomial distribution to generate depth counts}

\item{disper.par}{dispersion parameter for negative binomial distribution}

\item{chr.mb}{Chromossome size in mega base.}

\item{method}{Choose negative binomial ("neg.binom") distribution or updog ("updog") model to simulate counts values}

\item{mean.phred}{Sequencing error parameter}

\item{pos}{Phisical map position of each marker}

\item{haplo.ref}{The homologue or haplotype which correspond to the reference genome}

\item{chr}{Chromosome where the marker is positioned}

\item{phase}{if TRUE the genotypes in VCF will be phased}
}
\value{
vcf file located in out.file defined path
}
\description{
Given PedigreeSim .dat .map and .chrom files generate vcf file with depths count
in AD format field estimated by negative binomial. Works only for biallelic markers, 
the allels can be named as A (reference) and B, or by DNA bases A, T, C and G. In this last case, 
a vector indicanting the reference allele is needed.
}
\references{
Voorrips, R. E. , Maliepaard, C. A. (2012) The simulation of meiosis in 
diploid and tetraploid organisms using various genetic models.
\emph{BMC Bioinformatics} 13  - ISSN 1471-2105 - 12 p.

Bilton,T. P. , Schofield,  M. R. , Black,  M. A. , Chagné, D. , Wilcox, P. L. 
and Dodds, K. G. (2018) Accounting for Errors in Low Coverage High-Throughput
Sequencing Data When Constructing Genetic Maps Using Biparental Outcrossed Populations.
\emph{Genetics}. ISSN 0016-6731.
}
\seealso{
vcf file description 
<http://www.internationalgenome.org/wiki/Analysis/Variant%20Call%20Format/vcf-variant-call-format-version-40/>
\code{\link{[onemap]{simu_updog}}}
}
\author{
Cristiane Taniguti, \email{chtaniguti@usp.br}
}
