% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_segregation.R
\name{print.onemap.segreg.test}
\alias{print.onemap.segreg.test}
\title{Show the results of segregation tests}
\usage{
\method{print}{onemap.segreg.test}(x, ...)
}
\arguments{
\item{x}{an object of class onemap.segreg.test}

\item{...}{currently ignored}
}
\value{
a dataframe with marker name, H0 hypothesis, chi-square statistics,
p-values, and % of individuals genotyped.
}
\description{
It shows the results of Chisquare tests performed for all markers in a onemap object
of class class bc.onemap, f2.onemap, riself.onemap, risib.onemap or outcross.
}
\examples{
data(example.out) # Loads a fake outcross dataset installed with onemap
Chi <- test_segregation(example.out) # Performs the chi-square test for all markers
print(Chi) # Shows the results

}

