% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_haplotypes.R
\name{progeny_haplotypes_counts}
\alias{progeny_haplotypes_counts}
\title{By now only for outcrossing}
\usage{
progeny_haplotypes_counts(x)
}
\arguments{
\item{x}{object of class onemap_progeny_haplotypes}
}
\description{
Genotypes with same probability for two genotypes will be removed
}
