% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_haplotypes.R
\name{progeny_haplotypes_counts}
\alias{progeny_haplotypes_counts}
\title{Generate graphic with the number of break points for each individual considering the most likely genotypes estimated by the HMM.
Genotypes with same probability for two genotypes are removed. By now, only available for outcrossing and f2 intercross.}
\usage{
progeny_haplotypes_counts(x)
}
\arguments{
\item{x}{object of class onemap_progeny_haplotypes}
}
\description{
Generate graphic with the number of break points for each individual considering the most likely genotypes estimated by the HMM.
Genotypes with same probability for two genotypes are removed. By now, only available for outcrossing and f2 intercross.
}
