% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.outcross.R
\name{read.outcross}
\alias{read.outcross}
\title{Read data from a full-sib progeny (outcrossing populations)}
\usage{
read.outcross(dir, file)
}
\arguments{
\item{dir}{directory where the input file is located.}

\item{file}{the name of the input file which contains the data to be read.}
}
\value{
An object of class \code{outcross}, i.e., a list with the following
components: \item{geno}{a matrix with integers indicating the genotypes
read for each marker. Each column contains data for a marker and each row
represents an individual.} \item{n.ind}{number of individuals.}
\item{n.mar}{number of markers.} \item{segr.type}{a vector with the
segregation type of each marker, as \code{strings}.} \item{segr.type.num}{a
vector with the segregation type of each marker, represented in a
simplified manner as integers, i.e. 1 corresponds to markers of type
\code{"A"}; 2 corresponds to markers of type \code{"B1.5"}; 3 corresponds
to markers of type \code{"B2.6"}; 4 corresponds to markers of type
\code{"B3.7"}; 5 corresponds to markers of type \code{"C.8"}; 6 corresponds
to markers of type \code{"D1"} and 7 corresponds to markers of type
\code{"D2"}} \item{n.phe}{the number of traits included in the file}
\item{pheno}{the name of the phenoytpes} \item{input}{the name of the input file.}
}
\description{
Imports data from a full-sib family derived from the cross of two outbred
parents and creates an object of class \code{outcross}.
}
\details{
The file format is quite similar to that used by \code{MAPMAKER/EXP}
(\cite{Lincoln et al.}, 1993). The first line contains three integers: the
number of individuals, the number of markers and the number of traits.

Next comes the genotype data for all markers. Each new marker is initiated
with a \dQuote{*} (without the quotes) followed by the marker name, without
any space between them. Each marker name is followed by the corresponding
segregation type, which may be: \code{"A.1"}, \code{"A.2"}, \code{"A.3"},
\code{"A.4"}, \code{"B1.5"}, \code{"B2.6"}, \code{"B3.7"}, \code{"C.8"},
\code{"D1.9"}, \code{"D1.10"}, \code{"D1.11"}, \code{"D1.12"},
\code{"D1.13"}, \code{"D2.14"}, \code{"D2.15"}, \code{"D2.16"},
\code{"D2.17"} or \code{"D2.18"} (without quotes) [see
\code{\link[onemap]{marker.type}} and \cite{Wu et al.} (2002) for details].

After the segregation type comes the genotype data for the
corresponding marker. Depending on the segregation type, genotypes may be
denoted by \code{ac}, \code{ad}, \code{bc}, \code{bd}, \code{a}, \code{ba},
\code{b}, \code{bc}, \code{ab} and \code{o}, in several possible
combinations. To make things easier, we have followed \strong{exactly} the
notation used by \cite{Wu et al.} (2002). Genotypes \emph{must} be
separated by commas. Missing values are denoted by \code{"-"}

Finally, if there is phenotypic data, it will be added just after the marker data.
They need to be separated by commas as well, using the same symbol for missing information.

The \code{example} directory in the package distribution contains an
example data file to be read with this function. Further instructions can
be found at the tutorial distributed along with the package.
}
\examples{

  \dontrun{
    outcr_data <-
read.outcross(dir="work_directory",file="data_file.txt")
  }

}
\author{
Adapted from Karl Broman (package \pkg{qtl}) by Gabriel R A
Margarido, \email{gramarga@gmail.com}, later with additions from Luciano C Silva
}
\references{
Broman, K. W., Wu, H., Churchill, G., Sen, S., Yandell, B.
(2008) \emph{qtl: Tools for analyzing QTL experiments} R package version
1.09-43

Lincoln, S. E., Daly, M. J. and Lander, E. S. (1993) Constructing genetic
linkage maps with MAPMAKER/EXP Version 3.0: a tutorial and reference
manual. \emph{A Whitehead Institute for Biomedical Research Technical
Report}.

Wu, R., Ma, C.-X., Painter, I. and Zeng, Z.-B. (2002) Simultaneous maximum
likelihood estimation of linkage and linkage phases in outcrossing species.
\emph{Theoretical Population Biology} 61: 349-363.
}
\seealso{
\code{example} directory in the package source.
}
\keyword{IO}

