% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_inds}
\alias{remove_inds}
\title{Remove individuals from the onemap object}
\usage{
remove_inds(onemap.obj = NULL, rm.ind = NULL, list.seqs = NULL)
}
\arguments{
\item{onemap.obj}{object of class onemap}

\item{rm.ind}{vector of characters with individuals names}

\item{list.seqs}{list of objects of class sequence}
}
\value{
An object of class \code{onemap} without the selected individuals
if onemap object is used as input, or a list of objects of class \code{sequence}
without the selected individuals if a list of sequences objects is use as input
}
\description{
Remove individuals from the onemap object
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
