% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suggest_lod.R
\name{suggest_lod}
\alias{suggest_lod}
\title{suggest_lod}
\usage{
suggest_lod(x)
}
\arguments{
\item{x}{an object of class bc.onemap, f2.onemap, riself.onemap, risib.onemap or outcross}
}
\value{
the suggested LOD to be used for testing linkage
}
\description{
It suggests a LOD Score for declaring statistical significance for two-point tests
for linkage between all pairs of markers, considering that multiple tests are being performed.
}
\details{
In a somehow naive approach, the function calculates the number of two-point tests that
will be performed for all markers in the data set, and then using this to calculate
the global alpha required to control type I error using Bonferroni's correction.

From this global alpha, the corresponding quantil from the chi-square distribution is taken
and then converted to LOD Score.

This can be seen as just an initial approximation to help users to select a LOD Score for two
point tests.
}
\examples{
data(fake.bc.onemap) # Loads a fake backcross dataset installed with onemap
suggest_lod(fake.bc.onemap) # An value that should be used to start the analysis
}

