% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_seq.R
\name{try_seq_by_seq}
\alias{try_seq_by_seq}
\title{It uses try_seq function repeatedly trying to positioned each marker in a vector of markers into a already ordered sequence.
Each marker in the vector \code{"markers"} is kept in the sequence if the difference of LOD and total group size of the models 
with and without the marker are below the thresholds \code{"lod.thr"} and \code{"cM.thr"}.}
\usage{
try_seq_by_seq(sequence, markers, cM.thr = 10, lod.thr = -10)
}
\arguments{
\item{sequence}{object of class sequence with ordered markers}

\item{markers}{vector of integers defining the marker numbers to be inserted in the \code{sequence}}

\item{cM.thr}{number defining the threshold for total map size increase when inserting a single marker}

\item{lod.thr}{the difference of LODs between model before and after inserting the marker need to have 
value higher than the value defined in this argument}
}
\description{
It uses try_seq function repeatedly trying to positioned each marker in a vector of markers into a already ordered sequence.
Each marker in the vector \code{"markers"} is kept in the sequence if the difference of LOD and total group size of the models 
with and without the marker are below the thresholds \code{"lod.thr"} and \code{"cM.thr"}.
}
