% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ug.R
\name{ug}
\alias{ug}
\title{Unidirectional Growth}
\usage{
ug(input.seq, LOD = 0, max.rf = 0.5, tol = 1e-04)
}
\arguments{
\item{input.seq}{an object of class \code{sequence}.}

\item{LOD}{minimum LOD-Score threshold used when constructing the pairwise
recombination fraction matrix.}

\item{max.rf}{maximum recombination fraction threshold used as the LOD
value above.}

\item{tol}{tolerance for the C routine, i.e., the value used to evaluate
convergence.}
}
\value{
An object of class \code{sequence}, which is a list containing the
following components: \item{seq.num}{a \code{vector} containing the
(ordered) indices of markers in the sequence, according to the input file.}
\item{seq.phases}{a \code{vector} with the linkage phases between markers
in the sequence, in corresponding positions. \code{-1} means taht there are
no defined linkage phases.} \item{seq.rf}{a \code{vector} with the
recombination frequencies between markers in the sequence. \code{-1} means
that there are no estimated recombination frequencies.}
\item{seq.like}{log-likelihood of the corresponding linkage map.}
\item{data.name}{name of the object of class \code{onemap} with the raw
data.} \item{twopt}{name of the object of class \code{rf.2pts} with the
2-point analyses.}
}
\description{
Implements the marker ordering algorithm \emph{Unidirectional Growth}
(\cite{Tan & Fu, 2006}).
}
\details{
\emph{Unidirectional Growth} (\emph{UG}) is an algorithm for marker
ordering in linkage groups. It is not an exhaustive search method and,
therefore, is not computationally intensive. However, it does not guarantee
that the best order is always found. The only requirement is a matrix with
recombination fractions between markers.

After determining the order with \emph{UG}, the final map is constructed
using the multipoint approach (function \code{\link[onemap]{map}}).
}
\examples{

\dontrun{
  #outcross example
  data(example.out)
  twopt <- rf.2pts(example.out)
  all.mark <- make.seq(twopt,"all")
  groups <- group(all.mark)
  LG1 <- make.seq(groups,1)
  LG1.ug <- ug(LG1)

  #F2 example
  data(fake.f2.onemap)
  twopt <- rf.2pts(fake.f2.onemap)
  all.mark <- make.seq(twopt,"all")
  groups <- group(all.mark)
  LG1 <- make.seq(groups,1)
  LG1.ug <- ug(LG1)
  LG1.ug
}

}
\author{
Marcelo Mollinari, \email{mmollina@usp.br}
}
\references{
Mollinari, M., Margarido, G. R. A., Vencovsky, R. and Garcia,
A. A. F. (2009) Evaluation of algorithms used to order markers on genetics
maps. \emph{Heredity} 103: 494-502.

Tan, Y. and Fu, Y. (2006) A novel method for estimating linkage maps.
\emph{Genetics} 173: 2383-2390.
}
\seealso{
\code{\link[onemap]{make.seq}}, \code{\link[onemap]{map}}
}
\keyword{utilities}

