% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2raw.R
\name{vcf2raw}
\alias{vcf2raw}
\title{These functions are defunct and no longer available.}
\usage{
vcf2raw()
}
\value{
No return value, called for side effects
}
\description{
These functions are defunct and no longer available.
}
